Remove the SkillSetupService pass-through from the Domain layer
Every SkillSetupService method forwarded 1:1 to ISkillSetupPort, and its
Debug.Assert / cancellation preconditions are duplicated byte-for-byte in
the sole port implementation (ToolSkillSetupService), so the wrapper added
no behavior. SkillSetupUseCase and the composition root now depend on
ISkillSetupPort directly. ISkillSetupPort and SkillSetupTargetInfo move
into their own Domain files, and the Domain-residency guard test is
repointed at the port interface.
