+++ Assets/Tests/Editor/OnionAssemblyDependencyTests.cs
@@ -241,11 +241,11 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
         public void SkillSetupPolicy_WhenLoaded_CompilesUnderDomainAssembly()
         {
             // Tests that skill setup targets and policy services stay in the domain layer.
-            string serviceAssemblyName = typeof(SkillSetupService).Assembly.GetName().Name;
+            string portAssemblyName = typeof(ISkillSetupPort).Assembly.GetName().Name;
             string targetAssemblyName = typeof(SkillSetupTargetInfo).Assembly.GetName().Name;
             string stateAssemblyName = typeof(SkillInstallState).Assembly.GetName().Name;
 
-            Assert.That(serviceAssemblyName, Is.EqualTo(DomainAssemblyName));
+            Assert.That(portAssemblyName, Is.EqualTo(DomainAssemblyName));
             Assert.That(targetAssemblyName, Is.EqualTo(DomainAssemblyName));
             Assert.That(stateAssemblyName, Is.EqualTo(DomainAssemblyName));
         }
