+++ Packages/src/Editor/Domain/ISkillSetupPort.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 6827749ac9d41434bb4f768bc6d50d1f
+guid: 55553eda9da5491e9fbaa31772b2512a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+  userData:
+  assetBundleName:
+  assetBundleVariant:
