+++ Packages/src/Editor/Domain/SkillSetupTargetInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e85612eabee34c90bcf3423f5058362f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
